Move log_level to util.__init__
